Add logic to CompanyRecordsServiceImpl, related updates across mapper, repos and test(in progress)

+++ src/main/java/org/acme/api/strategies/FacebookStrategy.java
@@ -8,7 +8,7 @@ import org.acme.dto.CompanyDto;
 public class FacebookStrategy implements CompanyStrategy{
 
     @Override
-    public List<CompanyDto> getCompanies(int alumniId) {
+    public List<CompanyDto> getCompanies(long alumniId) {
         return new ArrayList<CompanyDto>();
     }
